// back-end/models/interface/IDocumentación_y_mantenimiento.ts
export type IDocumentación_y_mantenimiento = {
  id?: number
  cédula_verde__azul__título_del_automotor: string
  vtv_vigente: string
  segundo_juego_de_llaves: string
  informe_de_dominio_histórico: string
  escaneo_computarizado: string
  verificación_policial_vigente: string
  manual_de_usuario_con_servicios_registrados: string
  funcionamiento_del_control_remoto: string
  libre_de_deudas_financieras__multas_y_patentes: string
}